// apps/desktop/src/contracts/generated/DiagnosticsReport.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CompanionDiagnostics } from "./CompanionDiagnostics";
import type { DesktopDiagnostics } from "./DesktopDiagnostics";

export type DiagnosticsReport = { desktop: DesktopDiagnostics, companion: CompanionDiagnostics | null, };