// apps/desktop/src/contracts/generated/IntegrationReport.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CompanionIntegration } from "./CompanionIntegration";
import type { ServiceReport } from "./ServiceReport";

export type IntegrationReport = { service: ServiceReport, companion: CompanionIntegration | null, };